# analyses/OpenAPI_ColisApp.yaml
openapi: 3.0.0
info:
  title: ColisApp API
  description: Documentation de l'API RESTful pour l'application de gestion des envois ColisApp.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Serveur de développement local
  - url: https://colisapp.example.com/api/v1
    description: Serveur de production
security:
  - BearerAuth: []
...
paths:
  /users:
    get:
      summary: Récupérer la liste des utilisateurs
      description: Récupère une liste d'utilisateurs en fonction du rôle de l'utilisateur authentifié. Les super administrateurs voient tous les utilisateurs, les administrateurs d'agence voient ceux associés à leur agence.
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste d'utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /users/{id}:
    get:
      summary: Récupérer un utilisateur par ID
      description: Récupère un enregistrement d'utilisateur unique par son ID. Nécessite une authentification.
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l'utilisateur à récupérer
      responses:
        '200':
          description: Détails de l'utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Supprimer un utilisateur par ID
      description: Supprime un enregistrement d'utilisateur par son ID (suppression logique).
      tags:
        - Users
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l'utilisateur à supprimer
      responses:
        '200':
          description: Utilisateur supprimé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserDto' # Assuming it returns the deleted user or a confirmation
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /envois/user/{userId}:
    get:
      summary: Récupérer les envois par ID utilisateur
      description: Récupère une liste d'envois pour un utilisateur spécifique. Accessible par SUPER_ADMIN et l'utilisateur lui-même.
      tags:
        - Envois
        - Users
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: ID de l'utilisateur
      responses:
        '200':
          description: Liste d'envois pour l'utilisateur
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvoisListDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payment/complete-payment:
    get:
      summary: Finaliser le processus de paiement
      description: Finalise le processus de paiement après la redirection depuis la passerelle de paiement. Met à jour le statut de l'envoi et supprime le cookie de simulation.
      tags:
        - Payments
      responses:
        '200':
          description: Paiement effectué avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payment:
    post:
      summary: Créer une session de paiement Stripe
      description: Crée une nouvelle session de paiement Stripe.
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
              required:
                - amount
      responses:
        '200':
          description: ID de session Stripe
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /tarifs:
    get:
      summary: Récupérer les tarifs
      description: Récupère les tarifs de tarification.
      tags:
        - Tarifs
      responses:
        '200':
          description: Liste des tarifs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TarifsDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /transports:
    get:
      summary: Récupérer la liste des transports
      description: Récupère une liste de tous les transports.
      tags:
        - Transports
      responses:
        '200':
          description: Transports récupérés avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransportResponseDto'
                  message:
                    type: string
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Mettre à jour un transport
      description: Met à jour un transport existant avec les données fournies.
      tags:
        - Transports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransportRequestDto'
      responses:
        '200':
          description: Transport mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TransportResponseDto'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Non autorisé
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Not found
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid data provided
    MethodNotAllowed:
      description: Méthode non autorisée
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Method not allowed
    InternalServerError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
    Forbidden:
      description: Accès refusé
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Forbidden
  schemas:
...
    RoleDto:
      type: string
      enum:
        - CLIENT
        - SUPER_ADMIN
        - DESTINATAIRE
        - AGENCY_ADMIN
        - ACCOUNTANT
...
    AddressType:
      type: string
      enum:
        - HOME
        - OFFICE
        - BILLING
        - SHIPPING
        - OTHER
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
        birthDate:
          type: string
          format: date
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
          format: password
        image:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/RoleDto'
        agencyId:
          type: integer
        isVerified:
          type: boolean
        emailVerified:
          type: string
          format: date-time
        verificationToken:
          type: string
        verificationTokenExpires:
          type: string
          format: date-time
        userAddresses:
          $ref: '#/components/schemas/UserAddressDto'
          nullable: true
      required:
        - email
    ProfileDto:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        name:
          type: string
        phoneNumber:
          type: string
        userAddresses:
          $ref: '#/components/schemas/UserAddressDto'
        image:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/RoleDto'
        isVerified:
          type: boolean
      required:
        - id
        - email
        - firstName
        - lastName
        - birthDate
        - name
        - phoneNumber
        - userAddresses
        - role
        - isVerified
...
    AddressDto:
      type: object
      properties:
        id:
          type: integer
        street:
          type: string
        complement:
          type: string
          nullable: true
        streetNumber:
          type: string
          nullable: true
        boxNumber:
          type: string
          nullable: true
        cityId:
          type: integer
        city:
          $ref: '#/components/schemas/CityDto'
      required:
        - street
        - city
    UserAddressDto:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        addressId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        addressType:
          $ref: '#/components/schemas/AddressType'
        isDeleted:
          type: boolean
        address:
          $ref: '#/components/schemas/AddressDto'
        user:
          $ref: '#/components/schemas/UserDto'
        city:
          $ref: '#/components/schemas/CityDto'
      required:
        - userId
        - addressId
        - createdAt
        - updatedAt
        - addressType
        - isDeleted
        - address
        - user
        - city
...
    CityDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          $ref: '#/components/schemas/CountryDto'
      required:
        - id
        - name
        - country

    # Country DTOs
    CountryDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        iso2:
          type: string
          nullable: true
        iso3:
          type: string
          nullable: true
        phonecode:
          type: string
          nullable: true
        capital:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
      required:
        - id
        - name