# src/app/api/v1/payment/complete-payment/route.ts
// Path: src/app/api/v1/payment/complete-payment/route.ts
import { NextRequest, NextResponse } from "next/server";
import { completePaymentService } from "@/services/frontend-services/payment/paymentService";
import {
  getSimulationById,
  updatePaidEnvoi,
} from "@/services/backend-services/Bk_SimulationService";
import { verifySimulationToken } from "@/utils/verifySimulationToken";
import { EnvoiStatusDto, SimulationStatusDto } from "@/services/dtos";
import { cookies } from "next/headers";

/**
 * Complete payment process
 * @description Finalizes the payment process after redirection from the payment gateway. Updates envoi status and deletes the simulation cookie.
 * @response 200:{ message: string }:Payment completed successfully
 * @response 500:{ message: string }:Payment error
 * @openapi
 */
export async function GET(req: NextRequest) {
  try {
    // 1. Retrieve Simulation Data from Cookie
    const simulationFromCookie = verifySimulationToken(req);

    if (!simulationFromCookie) {
      return NextResponse.json({ data: null }, { status: 200 });
    }

    let simulation = await getSimulationById(Number(simulationFromCookie.id));

    if (!simulation) {
      return NextResponse.json({ data: null }, { status: 200 });
    }
    // 2. Update Simulation Status to COMPLETED and STATUS to PENDING and PAID to TRUE
    simulation =
      simulation.simulationStatus === SimulationStatusDto.CONFIRMED
        ? {
            ...simulation,
            simulationStatus: SimulationStatusDto.COMPLETED,
            envoiStatus: EnvoiStatusDto.PENDING,
          }
        : simulation;

    if (!simulation) {
      return NextResponse.json({ data: null }, { status: 200 });
    }

    // Update Simulation envoi status and simulationStatus
    await updatePaidEnvoi(simulation, simulationFromCookie);

    await completePaymentService();

    const response = NextResponse.json({
      message: "Payment completed successfully",
    });

    // delete cookies of the simulation and redirect to the profile page
    const cookieName = process.env.SIMULATION_COOKIE_NAME;

    // Actually remove the cookie from the response
    response.cookies.set({
      name: "cookieName", // Replace with your cookie name
      value: "",
      path: "/",
      maxAge: 0,
      httpOnly: true,
      sameSite: "lax",
    });

    return response;
  } catch (error) {
    console.error("Erreur lors de la création de la session Stripe:", error);
    return new NextResponse("Erreur de paiement", { status: 500 });
  }
}

export const dynamic = "force-dynamic";